requirements_celeb.txt
# Deep Learning
tensorflow>=2.12.0
keras>=2.12.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0
notebook>=7.0.0

# Optional (for production)
gunicorn>=21.2.0
python-dotenv>=1.0.0